refactor: promise then 패턴을 async await로 변경
가독성을 위해 비동기 로직 변경 및 파일 위치 수정
